Restructure methods to isolate responsibilities
