+++ cogs/stats_commands.py
@@ -2,10 +2,11 @@ import discord
 from discord.ext import commands
 from random import choice, randrange
 
-from utils.messages import embedded_stats, embedded_matchday_results, embed_dict
+from utils.messages import embedded_stats, embedded_matchday_results
+from utils.embedder import dict_to_embed
 from utils.dummy_data import get_dummy_data
-from the_mines.process.fussballdaten.process_html import get_matchday_results, get_blurb
-
+from the_mines.process.fussballdaten.process_matchday import get_matchday_results
+from the_mines.process.fussballdaten.process_blurb import get_blurb
 
 class Stats(commands.Cog):
     """All statistcal commands are collected here
@@ -51,7 +52,7 @@ class Stats(commands.Cog):
     @commands.command()
     async def blurb(self, ctx, team, season):
         team_name, blurb = get_blurb(team, season)
-        embed = embed_dict(ctx.author.display_name, ctx.author.avatar_url, team_name, blurb)
+        embed = dict_to_embed(ctx.author.display_name, ctx.author.avatar_url, team_name, blurb)
         await ctx.send(embed=embed)
